requirements.txt
requests==2.32.3
python-dotenv==1.0.1
Pillow==10.4.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
